requirements.txt
motor==1.1
numpy==1.14.1
pycparser==2.18
PyJWT==1.5.3
pymongo==3.5.1
python-dateutil==2.6.1
pytz==2018.3
requests<=2.18.4
tornado==4.5.2
urllib3==1.22
git+https://github.com/newwaybrazil/mongo-odm.git